+++ src/quantbox/plugins/strategies/__init__.py
@@ -8,6 +8,7 @@ from .crypto_trend import CryptoTrendStrategy
 from .ml_strategy import MLPredictionStrategy
 from .momentum_long_short import MomentumLongShortStrategy
 from .portfolio_optimizer import PortfolioOptimizerStrategy
+from .static_weights import StaticWeightsStrategy
 from .weighted_avg_aggregator import WeightedAverageAggregator
 
 __all__ = [
@@ -21,6 +22,7 @@ __all__ = [
     "CryptoRegimeTrendStrategy",
     "MLPredictionStrategy",
     "PortfolioOptimizerStrategy",
+    "StaticWeightsStrategy",
     "WeightedAverageAggregator",
     "select_universe",
     "select_universe_duckdb",